add version specifiers to pandas and python requirements
This fixes a linter complaint ("noarch: python recipes are required to have a lower bound on the python version") that's showing up in PRs like https://github.com/conda-forge/fredapi-feedstock/pull/6.

I'm not sure if the python 3.2 lower bound is actually correct. The fredapi package's contents itself do not specify a minimum Python version in the documentation or metadata, as far as I can see. I selected 3.2 because that's the minimum version required by pandas 0.15, and the fredapi package does declare a 'pandas>=0.15' dependency in its requirements.txt.